Added tests to check loading lost care modules when enabled/disabled
